+++ certbot/_internal/plugins/disco.py
