+++ certbot/_internal/storage.py
@@ -21,7 +21,7 @@ from certbot import util
 from certbot.compat import os
 from certbot.compat import filesystem
 from certbot.plugins import common as plugins_common
-from certbot.plugins import disco as plugins_disco
+from certbot._internal.plugins import disco as plugins_disco
 
 logger = logging.getLogger(__name__)
