+++ certbot/plugins/selection_test.py
@@ -11,40 +11,40 @@ from certbot import errors
 from certbot import interfaces
 from certbot.compat import os
 from certbot.display import util as display_util
-from certbot.plugins.disco import PluginsRegistry
+from certbot._internal.plugins.disco import PluginsRegistry
 from certbot.tests import util as test_util
 
 
 class ConveniencePickPluginTest(unittest.TestCase):
-    """Tests for certbot.plugins.selection.pick_*."""
+    """Tests for certbot._internal.plugins.selection.pick_*."""
 
     def _test(self, fun, ifaces):
         config = mock.Mock()
         default = mock.Mock()
         plugins = mock.Mock()
 
-        with mock.patch("certbot.plugins.selection.pick_plugin") as mock_p:
+        with mock.patch("certbot._internal.plugins.selection.pick_plugin") as mock_p:
             mock_p.return_value = "foo"
             self.assertEqual("foo", fun(config, default, plugins, "Question?"))
         mock_p.assert_called_once_with(
             config, default, plugins, "Question?", ifaces)
 
     def test_authenticator(self):
-        from certbot.plugins.selection import pick_authenticator
+        from certbot._internal.plugins.selection import pick_authenticator
         self._test(pick_authenticator, (interfaces.IAuthenticator,))
 
     def test_installer(self):
-        from certbot.plugins.selection import pick_installer
+        from certbot._internal.plugins.selection import pick_installer
         self._test(pick_installer, (interfaces.IInstaller,))
 
     def test_configurator(self):
-        from certbot.plugins.selection import pick_configurator
+        from certbot._internal.plugins.selection import pick_configurator
         self._test(pick_configurator,
             (interfaces.IAuthenticator, interfaces.IInstaller))
 
 
 class PickPluginTest(unittest.TestCase):
-    """Tests for certbot.plugins.selection.pick_plugin."""
+    """Tests for certbot._internal.plugins.selection.pick_plugin."""
 
     def setUp(self):
         self.config = mock.Mock(noninteractive_mode=False)
@@ -54,7 +54,7 @@ class PickPluginTest(unittest.TestCase):
         self.ifaces = []  # type: List[interfaces.IPlugin]
 
     def _call(self):
-        from certbot.plugins.selection import pick_plugin
+        from certbot._internal.plugins.selection import pick_plugin
         return pick_plugin(self.config, self.default, self.reg,
                            self.question, self.ifaces)
 
@@ -95,7 +95,7 @@ class PickPluginTest(unittest.TestCase):
             "bar": plugin_ep,
             "baz": plugin_ep,
         }
-        with mock.patch("certbot.plugins.selection.choose_plugin") as mock_choose:
+        with mock.patch("certbot._internal.plugins.selection.choose_plugin") as mock_choose:
             mock_choose.return_value = plugin_ep
             self.assertEqual("foo", self._call())
         mock_choose.assert_called_once_with(
@@ -107,13 +107,13 @@ class PickPluginTest(unittest.TestCase):
             "baz": None,
         }
 
-        with mock.patch("certbot.plugins.selection.choose_plugin") as mock_choose:
+        with mock.patch("certbot._internal.plugins.selection.choose_plugin") as mock_choose:
             mock_choose.return_value = None
             self.assertTrue(self._call() is None)
 
 
 class ChoosePluginTest(unittest.TestCase):
-    """Tests for certbot.plugins.selection.choose_plugin."""
+    """Tests for certbot._internal.plugins.selection.choose_plugin."""
 
     def setUp(self):
         zope.component.provideUtility(display_util.FileDisplay(sys.stdout,
@@ -130,17 +130,17 @@ class ChoosePluginTest(unittest.TestCase):
         ]
 
     def _call(self):
-        from certbot.plugins.selection import choose_plugin
+        from certbot._internal.plugins.selection import choose_plugin
         return choose_plugin(self.plugins, "Question?")
 
-    @test_util.patch_get_utility("certbot.plugins.selection.z_util")
+    @test_util.patch_get_utility("certbot._internal.plugins.selection.z_util")
     def test_selection(self, mock_util):
         mock_util().menu.side_effect = [(display_util.OK, 0),
                                         (display_util.OK, 1)]
         self.assertEqual(self.mock_stand, self._call())
         self.assertEqual(mock_util().notification.call_count, 1)
 
-    @test_util.patch_get_utility("certbot.plugins.selection.z_util")
+    @test_util.patch_get_utility("certbot._internal.plugins.selection.z_util")
     def test_more_info(self, mock_util):
         mock_util().menu.side_effect = [
             (display_util.OK, 1),
@@ -148,12 +148,12 @@ class ChoosePluginTest(unittest.TestCase):
 
         self.assertEqual(self.mock_stand, self._call())
 
-    @test_util.patch_get_utility("certbot.plugins.selection.z_util")
+    @test_util.patch_get_utility("certbot._internal.plugins.selection.z_util")
     def test_no_choice(self, mock_util):
         mock_util().menu.return_value = (display_util.CANCEL, 0)
         self.assertTrue(self._call() is None)
 
-    @test_util.patch_get_utility("certbot.plugins.selection.z_util")
+    @test_util.patch_get_utility("certbot._internal.plugins.selection.z_util")
     def test_new_interaction_avoidance(self, mock_util):
         mock_nginx = mock.Mock(
             description_with_name="n", misconfigured=False)
@@ -174,7 +174,7 @@ class ChoosePluginTest(unittest.TestCase):
         self.assertTrue("default" in mock_util().menu.call_args[1])
 
 class GetUnpreparedInstallerTest(test_util.ConfigTestCase):
-    """Tests for certbot.plugins.selection.get_unprepared_installer."""
+    """Tests for certbot._internal.plugins.selection.get_unprepared_installer."""
 
     def setUp(self):
         super(GetUnpreparedInstallerTest, self).setUp()
@@ -192,7 +192,7 @@ class GetUnpreparedInstallerTest(test_util.ConfigTestCase):
         })
 
     def _call(self):
-        from certbot.plugins.selection import get_unprepared_installer
+        from certbot._internal.plugins.selection import get_unprepared_installer
         return get_unprepared_installer(self.config, self.plugins)
 
     def test_no_installer_defined(self):
