--- docs/api/plugins/manual.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.plugins.manual`
----------------------------------
-
-.. automodule:: certbot.plugins.manual
-   :members:
