--- docs/api/plugins/selection.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.plugins.selection`
-------------------------------------
-
-.. automodule:: certbot.plugins.selection
-   :members:
